Serial # records now carry meta data about the cert * revoked - if toggled to 1, it means the serial number has been revoked * expiration - the expiration date of the associated certificate * collected - if the serial number record has been placed in the CRL
Both identity and entitlement certificates reference Serial # records

+++ proxy/src/main/java/org/fedoraproject/candlepin/model/CertificateSerialDto.java
@@ -14,8 +14,6 @@
  */
 package org.fedoraproject.candlepin.model;
 
-import java.math.BigInteger;
-
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlRootElement;
@@ -28,20 +26,20 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlAccessorType(XmlAccessType.PROPERTY)
 public class CertificateSerialDto {
 
-    private BigInteger serial;
+    private Long serial;
 
     public CertificateSerialDto() {
     }
 
-    public CertificateSerialDto(BigInteger serial) {
+    public CertificateSerialDto(Long serial) {
         this.serial = serial;
     }
 
-    public BigInteger getSerial() {
+    public Long getSerial() {
         return serial;
     }
 
-    public void setSerial(BigInteger serial) {
+    public void setSerial(Long serial) {
         this.serial = serial;
     }
